[DUT-845]: RegisterWard FSD 1차 분리
